Moving incremental tests to individual file
